Create actors in the plot
There is a TweetingCake which is being cut by a CakeSlicer. The
CakeSlicer alerts that is has served a piece of cake to twitter.
There is a supervisor who asks the user for her twitter handle and
after that supervises when the cake has been cut, so the CakeSlicer
knows when the slice has been served